pyproject.toml
[dependency-groups]
dev = [
  "Werkzeug[watchdog]",
  "ipdb",
  "pytest",
  "pytest-sugar",
  "coverage",
  "pre-commit",
  "django-debug-toolbar",
  "django-extensions",
  "django-coverage-plugin",
  "pytest-django",
  "django-models2puml>=0.2",
  "playwright",
  "pytest-playwright",
  "django-stubs[compatible-mypy]",
  "mypy",
  "djangorestframework-stubs[compatible-mypy]"
]
prod = ["gunicorn", "sentry-sdk>=1.40.5"]

[project]
authors = [{name = "Niccolò Cantù", email = "niccolo.cantu@nina.no"}]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
dependencies = [
  "django-environ",
  "python-slugify>=8.0.1",
  "Pillow>=10.3.0",
  "rcssmin",
  "argon2-cffi",
  "psycopg[binary]",
  "backoff",
  "requests",
  "django>=5.2.0",
  "django-model-utils",
  "django-allauth==65.4",
  "django-crispy-forms",
  "crispy-bootstrap5",
  "django-health-check",
  "django-probes",
  "django-postgres-extra",
  "django-solo",
  "django-appconf>=1.0.6",
  "django-cors-headers>=4.3.1",
  "django-filter>=24.1",
  "slippers",
  "django-tables2>=2.7.0",
  "django-autocomplete-light>=3.11.0",
  "dj-rest-auth",
  "django-rest-framework",
  "drf-spectacular",
  "drf-standardized-errors[openapi]",
  "djangorestframework-simplejwt",
  "django-cors-headers",
  "procrastinate[django]",
  "django-polymorphic",
  "django-unfold",
  "crispy-tailwind",
  "django-tailwind",
  "django-htmx",
  "django-template-partials",
  "neapolitan",
  "fontawesomefree",
  "django-allauth-ui==1.1.6",
  "django-widget-tweaks",
  "django-taggit",
  "django-tables2",
  "django-formset==1.7.6",
  "django-vite",
  "sqlglot[rs]",
  "django-view-breadcrumbs",
  "django-autocomplete-light",
  "djangorestframework-csv",
  "django-lifecycle>=1.2.4",
  "django-sequencefield>=1.0.13"
]
description = ""
license = {text = "GPL-3.0+"}
name = "genlab_bestilling"
readme = "README.md"
requires-python = ">=3.10"
version = "0.1.0"

# ==== Coverage ====
[tool.coverage.run]
include = ["app/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

[tool.django-stubs]
django_settings_module = "config.settings.local"

[tool.mypy]
disable_error_code = [
  "override"  # Python is a dynamic language, subclasses will get overwritten.
]
explicit_package_bases = true  # Use mypy_path only (ignores path passed when calling mypy).
follow_imports = "normal"
ignore_missing_imports = true  # Ignore missing imports, e.g. third-party packages without stubs. # Also covered by ruff.
local_partial_types = true
mypy_path = ["src"]  # Starting point when calling mypy. # https://mypy.readthedocs.io/en/stable/running_mypy.html#how-imports-are-found
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
warn_return_any = false
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
disable_error_code = ["var-annotated"]
module = "*.models"

# ==== pytest ====
[tool.pytest.ini_options]
addopts = "--ds=config.settings.test --reuse-db"
minversion = "6.0"
python_files = ["tests.py", "test_*.py"]

[tool.ruff]
fix = true

[tool.ruff.lint]
ignore = [
  "COM812",
  "RUF012",
  "RUF005",
  "RUF010",
  "ANN002",  # Too strict for introduction of ANN. # https://docs.astral.sh/ruff/rules/missing-type-args/
  "ANN003",  # Too strict for introduction of ANN. # https://docs.astral.sh/ruff/rules/missing-type-kwargs/
  "ANN204",  # Too strict for introduction of ANN. # https://docs.astral.sh/ruff/rules/missing-return-type-special-method/
  "ANN401",  # Too strict and too many errors atm. Consider enabling later after more typing is added. Search for `Any` to find unknown types. # https://docs.astral.sh/ruff/rules/any-type/
  "TRY003",  # This doesn't seem too bad. # https://docs.astral.sh/ruff/rules/raise-vanilla-args/
  "DJ001"  # Nullable string fields does is used a lot in this project. Maybe enable later? # https://docs.astral.sh/ruff/rules/django-nullable-model-string-field/
]
select = ["E", "W", "I", "F", "UP", "S", "B", "A", "COM", "LOG", "PTH", "Q", "PL", "RUF", "ANN", "PIE", "TC", "TRY", "DJ", "EM", "ISC", "RET", "SIM", "PERF"]

[tool.ruff.lint.flake8-annotations]
mypy-init-return = true  # https://docs.astral.sh/ruff/settings/#lint_flake8-annotations_mypy-init-return

[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["E501", "ANN", "I"]
"**/tests/*.py" = [
  "S101",
  "PLR2004",
  "ANN"  # Types on tests is more annoying than useful, types can be optional. # https://docs.astral.sh/ruff/rules/missing-return-type-undocumented-public-function/
]

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true
